requirements.txt
# 다중 웹사이트 RPA 시스템 의존성

# 핵심 라이브러리
setuptools>=68.0.0
wheel>=0.40.0
numpy
pandas
selenium>=4.15.0
openpyxl>=3.1.0
webdriver-manager>=4.0.0
pyyaml>=6.0.0
python-dotenv>=1.0.0
loguru>=0.7.0
requests>=2.31.0
beautifulsoup4>=4.12.0

# 확장 기능
schedule>=1.2.0
jinja2>=3.1.2
click>=8.1.3
rich>=13.3.0

# 개발 도구
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.11.0
flake8>=6.1.0

# 운영 도구
psutil>=5.9.5 